Move PageHero into its own file
